1、小白盒支付 2、打印

+++ app/src/main/java/com/huojumu/model/OrderDetails.java
@@ -84,7 +84,7 @@ public class OrderDetails {
         private String ordNo;
         private String payType;
         private String orderid;
-        private int totalPrice;
+        private double totalPrice;
         private String createTime;
         private String ordSource;
         private List<ProsBean> pros;
@@ -113,11 +113,11 @@ public class OrderDetails {
             this.orderid = orderid;
         }
 
-        public int getTotalPrice() {
+        public double getTotalPrice() {
             return totalPrice;
         }
 
-        public void setTotalPrice(int totalPrice) {
+        public void setTotalPrice(double totalPrice) {
             this.totalPrice = totalPrice;
         }
 
@@ -159,7 +159,7 @@ public class OrderDetails {
              */
 
             private String sumPrice;
-            private int price;
+            private double price;
             private int proId;
             private int proCount;
             private String proName;
@@ -176,11 +176,11 @@ public class OrderDetails {
                 this.sumPrice = sumPrice;
             }
 
-            public int getPrice() {
+            public double getPrice() {
                 return price;
             }
 
-            public void setPrice(int price) {
+            public void setPrice(double price) {
                 this.price = price;
             }
